DEV: Use the velocity to generate discrete trajectory
This method is slightly slower than the current derivative-less
method. The number of discrete points for my test was 8% higher
and the number of calls to triangle was also 5% higher. Maybe
the sampling distribution is more regular, but derivative
calculation is expensive even though we have an analytical
derivative function.