DEV: Tried preallocating points along discrete_trajectory
It takes mroe than twice as long because pre-allocating causes
four times as many points to be added to the trajectory.
Probably, if the time step is not an appropriate multiple of the
space step, then extra steps need to be generated to ensure that
the space step is appropriately small.